requirements.txt
Flask==3.1.1
requests==2.31.0
openai==1.66.0
anthropic==0.46.0
Werkzeug==3.1.3
boto3==1.35.96
python-dotenv==1.0.0